Se divide la logica de los traits
