+++ resources/views/livewire/calculadora-interes-simple.blade.php
@@ -1,5 +1,5 @@
 <div>
-    <form class="space-y-8" wire:submit.prevent="calcular('interesSimple')">
+    <form class="space-y-8" wire:submit.prevent="calcular()">
 
         <!-- Selector de fórmula -->
         <div class="bg-blue-50 dark:bg-blue-900/20 border border-blue-200 dark:border-blue-800 rounded-lg p-6 mb-6">
